Creaate directory for component
